Move `SqlCteNode` to a separate file.
