+++ metricflow/sql/optimizer/cte_mapping_lookup_builder.py
@@ -9,9 +9,8 @@ from typing_extensions import override
 
 from metricflow.sql.optimizer.cte_alias_to_cte_node_mapping import SqlCteAliasMappingLookup
 from metricflow.sql.sql_ctas_node import SqlCreateTableAsNode
+from metricflow.sql.sql_cte_node import SqlCteAliasMapping, SqlCteNode
 from metricflow.sql.sql_plan import (
-    SqlCteAliasMapping,
-    SqlCteNode,
     SqlPlanNode,
     SqlPlanNodeVisitor,
 )
